Correção de erro na injeção de dependência.

+++ EmergencyManagementSystem.Common.Entities/Entities/Employee.cs
@@ -18,6 +18,6 @@ namespace EmergencyManagementSystem.Common.Entities.Entities
         public virtual Address Address { get; set; }
         public long AddressId { get; set; }
         public virtual Occupation Occupation { get; set; }
-        public long OccupationId { get; set; }
+        public short OccupationId { get; set; }
     }
 }
